Week 4/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.4.0
boto3==1.42.31
botocore==1.42.31
brotli==1.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
comm==0.2.3
contourpy==1.3.0
cryptography==46.0.1
cycler==0.12.1
dataclasses-json==0.6.7
ddgs==9.8.0
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docx2txt==0.9
duckduckgo_search==8.1.1
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.128.0
filelock==3.19.1
filetype==1.2.0
fonttools==4.60.0
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.10
imbalanced-learn==0.12.4
importlib_resources==6.5.2
ipython==8.18.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
langchain==0.3.27
langchain-aws==0.2.35
langchain-community==0.3.31
langchain-core==0.3.83
langchain-google-genai==2.0.10
langchain-openai==0.3.35
langchain-pinecone==0.2.13
langchain-text-splitters==0.3.11
langsmith==0.4.37
lightgbm==4.6.0
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.3.1
openai==2.15.0
orjson==3.11.5
packaging==24.2
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
primp==0.15.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==6.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==5.1.2
simsimd==6.5.12
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
tenacity==9.1.2
teradataml==20.0.0.7
teradatamlwidgets==20.0.0.6
teradatasql==20.0.0.41
teradatasqlalchemy==20.0.0.7
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.8.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.6
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
uuid_utils==0.14.0
uvicorn==0.39.0
wcwidth==0.2.14
widgetsnbextension==4.0.14
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
